+++ src/app/find/password/layout.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const layout = () => {
+  return <div>123</div>;
+};
+
+export default layout;
