+++ src/app/find/password/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Page = () => {
+  return <div>비밀번호찾기페이지입니다</div>;
+};
+
+export default Page;
